fix: indent and rename to modify inspecteur from a reservqtion
