+++ client/src/store/index.js
@@ -4,7 +4,7 @@ import Vuex from 'vuex'
 
 import aurige from './aurige'
 import admin from './admin'
-import adminModifIpcsr from './admin-modif-ipcsr'
+import adminModifInspecteur from './admin-modif-inspecteur'
 import adminSearch from './admin-search'
 import auth from './auth'
 import candidat from './candidat'
@@ -18,7 +18,7 @@ import whitelist from './whitelist'
 
 export * from './aurige'
 export * from './admin'
-export * from './admin-modif-ipcsr'
+export * from './admin-modif-inspecteur'
 export * from './admin-search'
 export * from './auth'
 export * from './candidat'
@@ -37,7 +37,7 @@ export default new Vuex.Store({
   },
 
   modules: {
-    adminModifIpcsr,
+    adminModifInspecteur,
     adminSearch,
     aurige,
     admin,
